refactor: extract LockMixin from NexusFS
Move all lock methods (sys_lock, sys_unlock, lock_info, lock_list,
lock_extend, lock_force_release, lock_release, _acquire_lock_sync,
_release_lock_sync) to nexus_fs_lock.py as LockMixin.

~200 lines extracted. Same mixin pattern as WatchMixin.
